+++ src/PlaywrightSharp.Tests/Page/AsyncStacksTests.cs
@@ -1,0 +1,34 @@
+using System.Threading.Tasks;
+using PlaywrightSharp.Tests.BaseTests;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace PlaywrightSharp.Tests.Page
+{
+    ///<playwright-file>page.spec.js</playwright-file>
+    ///<playwright-describe>Async stacks</playwright-describe>
+    public class AsyncStacksTests : PlaywrightSharpPageBaseTest
+    {
+        internal AsyncStacksTests(ITestOutputHelper output) : base(output)
+        {
+        }
+
+        ///<playwright-file>page.spec.js</playwright-file>
+        ///<playwright-describe>Async stacks</playwright-describe>
+        ///<playwright-it>should work</playwright-it>
+        [Fact]
+        public async Task ShouldWork()
+        {
+            Server.SetRoute("/empty.html", context =>
+            {
+                context.Abort(); // is this right?
+                return Task.CompletedTask;
+            });
+            var exception = await Assert.ThrowsAsync<PlaywrightSharpException>(() => Page.GoToAsync(TestConstants.EmptyPage));
+            // TODO: what should we assert here?
+            // expect(error.stack).toContain(__filename);
+        }
+
+    }
+
+}
