+++ src/PlaywrightSharp.Tests/Page/PageCloseTests.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using PlaywrightSharp.Tests.Attributes;
@@ -122,53 +120,6 @@ namespace PlaywrightSharp.Tests.Page
             }
         }
     }
-
-    ///<playwright-file>page.spec.js</playwright-file>
-    ///<playwright-describe>Page.Events.Load</playwright-describe>
-    public class PageEventsLoadTests : PlaywrightSharpPageBaseTest
-    {
-        internal PageEventsLoadTests(ITestOutputHelper output) : base(output)
-        {
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.Load</playwright-describe>
-        ///<playwright-it>should fire when expected</playwright-it>
-        [Fact]
-        public async Task ShouldFireWhenExpected()
-        {
-            await Task.WhenAll(
-              Page.GoToAsync("about:blank"),
-              Page.WaitForEvent<EventArgs>(PageEvent.Load)
-            );
-        }
-    }
-
-    ///<playwright-file>page.spec.js</playwright-file>
-    ///<playwright-describe>Async stacks</playwright-describe>
-    public class AsyncStacksTests : PlaywrightSharpPageBaseTest
-    {
-        internal AsyncStacksTests(ITestOutputHelper output) : base(output)
-        {
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Async stacks</playwright-describe>
-        ///<playwright-it>should work</playwright-it>
-        [Fact]
-        public async Task ShouldWork()
-        {
-            Server.SetRoute("/empty.html", context =>
-            {
-                context.Abort(); // is this right?
-                return Task.CompletedTask;
-            });
-            var exception = await Assert.ThrowsAsync<PlaywrightSharpException>(() => Page.GoToAsync(TestConstants.EmptyPage));
-            // TODO: what should we assert here?
-            // expect(error.stack).toContain(__filename);
-        }
-
-    }
     ///<playwright-file>page.spec.js</playwright-file>
     ///<playwright-describe>Page.Events.error</playwright-describe>
     public class PageEventsErrorTests : PlaywrightSharpPageBaseTest
@@ -197,470 +148,6 @@ namespace PlaywrightSharp.Tests.Page
             await Page.WaitForEvent<ErrorEventArgs>(PageEvent.Error);
             Assert.Equal("Page crashed!", error);
         }
-    }
-
-    ///<playwright-file>page.spec.js</playwright-file>
-    ///<playwright-describe>Page.Events.Popup</playwright-describe>
-    public class PageEventsPopupTests : PlaywrightSharpPageBaseTest
-    {
-        internal PageEventsPopupTests(ITestOutputHelper output) : base(output)
-        {
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.Popup</playwright-describe>
-        ///<playwright-it>should work</playwright-it>
-        [Fact]
-        public async Task ShouldWork()
-        {
-            var popupTask = Page.OnceAsync<PopupEventArgs>(PageEvent.Popup);
-            await Task.WhenAll(
-                popupTask,
-                Page.EvaluateAsync("() => window.open('about:blank')")
-            );
-            var popup = popupTask.Result.Page;
-            Assert.False(await Page.EvaluateAsync<bool>("() => !!window.opener"));
-            Assert.True(await popup.EvaluateAsync<bool>("() => !!window.opener"));
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.Popup</playwright-describe>
-        ///<playwright-it>should work with noopener</playwright-it>
-        [Fact]
-        public async Task ShouldWorkWithNoopener()
-        {
-            var popupTask = Page.OnceAsync<PopupEventArgs>(PageEvent.Popup);
-            await Task.WhenAll(
-                popupTask,
-                Page.EvaluateAsync<string>("() => window.open('about:blank', null, 'noopener')")
-            );
-            var popup = popupTask.Result.Page;
-            Assert.False(await Page.EvaluateAsync<bool>("() => !!window.opener"));
-            Assert.False(await popup.EvaluateAsync<bool>("() => !!window.opener"));
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.Popup</playwright-describe>
-        ///<playwright-it>should work with clicking target=_blank</playwright-it>
-        [SkipBrowserAndPlatformFact(skipFirefox: true)]
-        public async Task ShouldWorkWithClickingTargetBlank()
-        {
-            await Page.GoToAsync(TestConstants.EmptyPage);
-            await Page.SetContentAsync("<a target=_blank rel=\"opener\" href=\"/one-style.html\">yo</a>");
-            var popupTask = Page.WaitForEvent<PopupEventArgs>(PageEvent.Popup).ContinueWith(async task =>
-            {
-                var popup = task.Result.Page;
-                await popup.WaitForLoadStateAsync();
-                return popup;
-            });
-            await Task.WhenAll(
-                await popupTask,
-                Page.ClickAsync("a")
-            );
-            var popup = await popupTask.Result;
-            Assert.False(await Page.EvaluateAsync<bool>("() => !!window.opener"));
-            Assert.True(await popup.EvaluateAsync<bool>("() => !!window.opener"));
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.Popup</playwright-describe>
-        ///<playwright-it>should work with fake-clicking target=_blank and rel=noopener</playwright-it>
-        [SkipBrowserAndPlatformFact(skipFirefox: true)]
-        public async Task ShouldWorkWithFakeClickingTargetBlankAndRelNoopener()
-        {
-            // TODO: FFOX sends events for "one-style.html" request to both pages.
-            await Page.GoToAsync(TestConstants.EmptyPage);
-            await Page.SetContentAsync("<a target=_blank rel=noopener href=\"/one-style.html\">yo</a>");
-            var popupTask = Page.WaitForEvent<PopupEventArgs>(PageEvent.Popup).ContinueWith(async task =>
-            {
-                var popup = task.Result.Page;
-                await popup.WaitForLoadStateAsync();
-                return popup;
-            });
-            await Task.WhenAll(
-                popupTask,
-                Page.QuerySelectorEvaluateAsync("a", "a => a.click()")
-            );
-            var popup = await popupTask.Result;
-            Assert.False(await Page.EvaluateAsync<bool>("() => !!window.opener"));
-            // TODO: At this point popup might still have about:blank as the current document.
-            // FFOX is slow enough to trigger this. We should do something about popups api.
-            Assert.False(await popup.EvaluateAsync<bool>("() => !!window.opener"));
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.Popup</playwright-describe>
-        ///<playwright-it>should work with clicking target=_blank and rel=noopener</playwright-it>
-        [SkipBrowserAndPlatformFact(skipFirefox: true)]
-        public async Task ShouldWorkWithClickingTargetBlankAndRelNoopener()
-        {
-            await Page.GoToAsync(TestConstants.EmptyPage);
-            await Page.SetContentAsync("<a target=_blank rel=noopener href=\"/one-style.html\">yo</a>");
-            var popupTask = Page.WaitForEvent<PopupEventArgs>(PageEvent.Popup).ContinueWith(async task =>
-            {
-                var popup = task.Result.Page;
-                await popup.WaitForLoadStateAsync();
-                return popup;
-            });
-            await Task.WhenAll(
-                popupTask,
-                Page.ClickAsync("a")
-            );
-            var popup = await popupTask.Result;
-            Assert.False(await Page.EvaluateAsync<bool>("() => !!window.opener"));
-            Assert.False(await popup.EvaluateAsync<bool>("() => !!window.opener"));
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.Popup</playwright-describe>
-        ///<playwright-it>should not treat navigations as new popups</playwright-it>
-        [SkipBrowserAndPlatformFact(skipFirefox: true)]
-        public async Task ShouldNotTreatNavigationsAsNewPopups()
-        {
-            await Page.GoToAsync(TestConstants.EmptyPage);
-            await Page.SetContentAsync("<a target=_blank rel=noopener href=\"/one-style.html\">yo</a>");
-            var popupTask = Page.WaitForEvent<PopupEventArgs>(PageEvent.Popup).ContinueWith(async task =>
-            {
-                var popup = task.Result.Page;
-                await popup.WaitForLoadStateAsync();
-                return popup;
-            });
-            await Task.WhenAll(
-                popupTask,
-                Page.ClickAsync("a")
-            );
-            var popup = await popupTask.Result;
-            bool badSecondPopup = false;
-            Page.Popup += (sender, e) => badSecondPopup = true;
-            await popup.GoToAsync(TestConstants.CrossProcessUrl + "/empty.html");
-            Assert.False(badSecondPopup);
-        }
-    }
-
-    ///<playwright-file>page.spec.js</playwright-file>
-    ///<playwright-describe>Page.opener</playwright-describe>
-    public class PageOpenerTests : PlaywrightSharpPageBaseTest
-    {
-        internal PageOpenerTests(ITestOutputHelper output) : base(output)
-        {
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.opener</playwright-describe>
-        ///<playwright-it>should provide access to the opener page</playwright-it>
-        [Fact]
-        public async Task ShouldProvideAccessToTheOpenerPage()
-        {
-            var popupTask = Page.OnceAsync<PopupEventArgs>(PageEvent.Popup);
-            await Task.WhenAll(
-                popupTask,
-                Page.EvaluateAsync("() => window.open('about:blank')")
-            );
-            var popup = popupTask.Result.Page;
-            var opener = await popup.GetOpenerAsync();
-            Assert.Equal(Page, opener);
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.opener</playwright-describe>
-        ///<playwright-it>should return null if parent page has been closed</playwright-it>
-        [Fact]
-        public async Task ShouldReturnNullIfParentPageHasBeenClosed()
-        {
-            var popupTask = Page.OnceAsync<PopupEventArgs>(PageEvent.Popup);
-            await Task.WhenAll(
-                popupTask,
-                Page.EvaluateAsync("() => window.open('about:blank')")
-            );
-            var popup = popupTask.Result.Page;
-            await Page.CloseAsync();
-            var opener = await popup.GetOpenerAsync();
-            Assert.Null(opener);
-        }
-    }
-
-    ///<playwright-file>page.spec.js</playwright-file>
-    ///<playwright-describe>Page.Events.Console</playwright-describe>
-    public class PageEventsConsoleTests : PlaywrightSharpPageBaseTest
-    {
-        internal PageEventsConsoleTests(ITestOutputHelper output) : base(output)
-        {
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.Console</playwright-describe>
-        ///<playwright-it>should work</playwright-it>
-        [Fact]
-        public async Task ShouldWork()
-        {
-            ConsoleMessage message = null;
-            Page.Once(PageEvent.Console, (ConsoleEventArgs e) => message = e.Message);
-            await Task.WhenAll(
-                Page.EvaluateAsync<string>("() => console.log('hello', 5, { foo: 'bar'})"),
-                Page.WaitForEvent<ConsoleEventArgs>(PageEvent.Console)
-            );
-            Assert.Equal("hello 5 JSHandle@object", message.Text);
-            Assert.Equal(ConsoleType.Log, message.Type);
-            Assert.Equal("hello", await message.Args[0].GetJsonValueAsync<string>());
-            Assert.Equal(5, await message.Args[1].GetJsonValueAsync<int>());
-            Assert.Equal(new { foo = "bar" }, await message.Args[2].GetJsonValueAsync<object>());
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.Console</playwright-describe>
-        ///<playwright-it>should work for different console API calls</playwright-it>
-        [Fact]
-        public async Task ShouldWorkForDifferentConsoleAPICalls()
-        {
-            var messages = new List<ConsoleMessage>();
-            Page.Console += (sender, e) => messages.Add(e.Message);
-            // All console events will be reported before `Page.evaluate` is finished.
-            await Page.EvaluateAsync(@"() => {
-                // A pair of time/timeEnd generates only one Console API call.
-                console.time('calling console.time');
-                console.timeEnd('calling console.time');
-                console.trace('calling console.trace');
-                console.dir('calling console.dir');
-                console.warn('calling console.warn');
-                console.error('calling console.error');
-                console.log(Promise.resolve('should not wait until resolved!'));
-            }");
-            Assert.Equal(new[] { ConsoleType.TimeEnd, ConsoleType.Trace, ConsoleType.Dir, ConsoleType.Warning, ConsoleType.Error, ConsoleType.Log }, messages.Select(msg => msg.Type));
-            Assert.Contains("calling console.time", messages[0].Text);
-            Assert.Equal(new[]
-            {
-                "calling console.trace",
-                "calling console.dir",
-                "calling console.warn",
-                "calling console.error",
-                "JSHandle@promise"
-            }, messages.Skip(1).Select(msg => msg.Text));
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.Console</playwright-describe>
-        ///<playwright-it>should not fail for window object</playwright-it>
-        [Fact]
-        public async Task ShouldNotFailForWindowObject()
-        {
-            ConsoleMessage message = null;
-            Page.Once(PageEvent.Console, (ConsoleEventArgs e) => message = e.Message);
-            await Task.WhenAll(
-                Page.EvaluateAsync("() => console.error(window)"),
-                Page.WaitForEvent<ConsoleEventArgs>(PageEvent.Console)
-            );
-            Assert.Equal("JSHandle@object", message.Text);
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.Console</playwright-describe>
-        ///<playwright-it>should trigger correct Log</playwright-it>
-        [Fact]
-        public async Task ShouldTriggerCorrectLog()
-        {
-            await Page.GoToAsync("about:blank");
-            var messageTask = Page.WaitForEvent<ConsoleEventArgs>(PageEvent.Console);
-            await Task.WhenAll(
-                messageTask,
-                Page.EvaluateAsync("async url => fetch(url).catch (e => { })", TestConstants.EmptyPage)
-            );
-            var message = messageTask.Result.Message;
-            Assert.Contains("Access-Control-Allow-Origin", message.Text);
-            Assert.Equal(ConsoleType.Error, message.Type);
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.Console</playwright-describe>
-        ///<playwright-it>should have location for console API calls</playwright-it>
-        [Fact]
-        public async Task ShouldHaveLocationForConsoleAPICalls()
-        {
-            await Page.GoToAsync(TestConstants.EmptyPage);
-            var messageTask = Page.WaitForEvent<ConsoleEventArgs>(PageEvent.Console);
-            await Task.WhenAll(
-                messageTask,
-                Page.GoToAsync(TestConstants.ServerUrl + "/consolelog.html")
-            );
-            var message = messageTask.Result.Message;
-            Assert.Equal("yellow", message.Text);
-            Assert.Equal(ConsoleType.Log, message.Type);
-            var location = message.Location;
-            // Engines have different column notion.
-            location.ColumnNumber = null;
-            Assert.Equal(new ConsoleMessageLocation
-            {
-                URL = TestConstants.ServerUrl + "/consolelog.html",
-                LineNumber = 7
-            }, location);
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.Console</playwright-describe>
-        ///<playwright-it>should not throw when there are console messages in detached iframes</playwright-it>
-        [SkipBrowserAndPlatformFact(skipFirefox: true)]
-        public async Task ShouldNotThrowWhenThereAreConsoleMessagesInDetachedIframes()
-        {
-            await Page.GoToAsync(TestConstants.EmptyPage);
-            await Page.EvaluateAsync(@"async () =>
-            {
-                // 1. Create a popup that Playwright is not connected to.
-                var win = window.open(window.location.href, 'Title', 'toolbar=no,location=no,directories=no,status=no,menubar=no,scrollbars=yes,resizable=yes,width=780,height=200,top=0,left=0');
-                while (window.document.readyState !== 'complete')
-                {
-                    await new Promise(f => setTimeout(f, 100));
-                }
-                // 2. In this popup, create an iframe that console.logs a message.
-                win.document.body.innerHTML = `< iframe src = '/consolelog.html' ></ iframe >`;
-                var frame = win.document.querySelector('iframe');
-                while (frame.contentDocument.readyState !== 'complete')
-                {
-                    await new Promise(f => setTimeout(f, 100));
-                }
-                // 3. After that, remove the iframe.
-                frame.remove();
-            }");
-            // 4. Connect to the popup and make sure it doesn't throw.
-            await Page.BrowserContext.GetPagesAsync();
-        }
-    }
-
-    ///<playwright-file>page.spec.js</playwright-file>
-    ///<playwright-describe>Page.Events.DOMContentLoaded</playwright-describe>
-    public class PageEventsDOMContentLoadedTests : PlaywrightSharpPageBaseTest
-    {
-        internal PageEventsDOMContentLoadedTests(ITestOutputHelper output) : base(output)
-        {
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.Events.DOMContentLoaded</playwright-describe>
-        ///<playwright-it>should fire when expected</playwright-it>
-        [Fact]
-        public async Task ShouldFireWhenExpected()
-        {
-            Page.GoToAsync("about:blank");
-            await Page.WaitForEvent<object>(PageEvent.DOMContentLoaded);
-        }
-    }
-
-    ///<playwright-file>page.spec.js</playwright-file>
-    ///<playwright-describe>Page.waitForRequest</playwright-describe>
-    public class PageWaitForRequestTests : PlaywrightSharpPageBaseTest
-    {
-        internal PageWaitForRequestTests(ITestOutputHelper output) : base(output)
-        {
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.waitForRequest</playwright-describe>
-        ///<playwright-it>should work</playwright-it>
-        [Fact]
-        public async Task ShouldWork()
-        {
-            await Page.GoToAsync(TestConstants.EmptyPage);
-            var requestTask = Page.WaitForRequestAsync(TestConstants.ServerUrl + "/digits/2.png");
-            await Task.WhenAll(
-                requestTask,
-                Page.EvaluateAsync(@"() => {
-                  fetch('/digits/1.png');
-                  fetch('/digits/2.png');
-                  fetch('/digits/3.png');
-                }")
-            );
-            var request = requestTask.Result;
-            Assert.Equal(TestConstants.ServerUrl + "/digits/2.png", request.Url);
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.waitForRequest</playwright-describe>
-        ///<playwright-it>should work with predicate</playwright-it>
-        [Fact]
-        public async Task ShouldWorkWithPredicate()
-        {
-            await Page.GoToAsync(TestConstants.EmptyPage);
-            var requestTask = Page.WaitForEvent(PageEvent.Request, new WaitForEventOptions<RequestEventArgs>
-            {
-                Predicate = e => e.Request.Url == TestConstants.ServerUrl + "/digits/2.png"
-            });
-            await Task.WhenAll(
-                requestTask,
-                Page.EvaluateAsync<string>(@"() => {
-                    fetch('/digits/1.png');
-                    fetch('/digits/2.png');
-                    fetch('/digits/3.png');
-                }")
-            );
-            var request = requestTask.Result;
-            Assert.Equal(TestConstants.ServerUrl + "/digits/2.png", request.Url);
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.waitForRequest</playwright-describe>
-        ///<playwright-it>should respect timeout</playwright-it>
-        [Fact]
-        public async Task ShouldRespectTimeout()
-        {
-            let error = null;
-            var exception = await Assert
-            await Page.waitForEvent('request', { predicate: () => false, timeout: 1 }).catch (e => error = e);
-            expect(error).toBeInstanceOf(playwright.errors.TimeoutError);
-
-            }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.waitForRequest</playwright-describe>
-        ///<playwright-it>should respect default timeout</playwright-it>
-        [Fact]
-        public async Task ShouldRespectDefaultTimeout()
-        {
-
-            let error = null;
-            Page.setDefaultTimeout(1);
-            await Page.waitForEvent('request', () => false).catch (e => error = e);
-            expect(error).toBeInstanceOf(playwright.errors.TimeoutError);
-
-            }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.waitForRequest</playwright-describe>
-        ///<playwright-it>should work with no timeout</playwright-it>
-        [Fact]
-        public async Task ShouldWorkWithNoTimeout()
-        {
-
-            await Page.GoToAsync(TestConstants.EmptyPage);
-            var[request] = await Promise.all([
-              Page.waitForRequest(TestConstants.ServerUrl + '/digits/2.png', { timeout: 0}),
-        Page.EvaluateAsync<string>(() => setTimeout(() =>
-        {
-            fetch('/digits/1.png');
-            fetch('/digits/2.png');
-            fetch('/digits/3.png');
-        }, 50))
-      ]);
-            expect(request.Url).toBe(TestConstants.ServerUrl + '/digits/2.png');
-
-        }
-
-        ///<playwright-file>page.spec.js</playwright-file>
-        ///<playwright-describe>Page.waitForRequest</playwright-describe>
-        ///<playwright-it>should work with url match</playwright-it>
-        [Fact]
-        public async Task ShouldWorkWithUrlMatch()
-        {
-
-            await Page.GoToAsync(TestConstants.EmptyPage);
-            var[request] = await Promise.all([
-              Page.waitForRequest(/ digits\/\d\.png /),
-              Page.EvaluateAsync<string>(() =>
-              {
-                  fetch('/digits/1.png');
-              })
-            ]);
-            expect(request.Url).toBe(TestConstants.ServerUrl + '/digits/1.png');
-
-        }
-
     }
     ///<playwright-file>page.spec.js</playwright-file>
     ///<playwright-describe>Page.waitForResponse</playwright-describe>
